Get rid of the split package in the gRPC extension artifacts
- resolves #21919
