+++ extensions/grpc/api/src/main/java/io/quarkus/grpc/GlobalInterceptor.java
