+++ extensions/grpc/api/src/main/java/io/quarkus/grpc/MutinyClient.java
@@ -1,4 +1,4 @@
-package io.quarkus.grpc.runtime;
+package io.quarkus.grpc;
 
 import io.grpc.stub.AbstractStub;
