+++ extensions/grpc/api/src/main/java/io/quarkus/grpc/MutinyService.java
@@ -1,4 +1,4 @@
-package io.quarkus.grpc.runtime;
+package io.quarkus.grpc;
 
 /**
  * Used to mark a Mutiny service interface generated for a gRPC service.
