+++ extensions/grpc/api/src/main/java/io/quarkus/grpc/MutinyStub.java
@@ -1,4 +1,4 @@
-package io.quarkus.grpc.runtime;
+package io.quarkus.grpc;
 
 /**
  * A marker interface that represents a generated Mutiny gRPC stub.
