+++ extensions/grpc/deployment/src/main/java/io/quarkus/grpc/deployment/GrpcDotNames.java
@@ -16,14 +16,14 @@ import io.quarkus.gizmo.MethodDescriptor;
 import io.quarkus.grpc.GlobalInterceptor;
 import io.quarkus.grpc.GrpcClient;
 import io.quarkus.grpc.GrpcService;
+import io.quarkus.grpc.MutinyBean;
+import io.quarkus.grpc.MutinyClient;
+import io.quarkus.grpc.MutinyGrpc;
+import io.quarkus.grpc.MutinyService;
+import io.quarkus.grpc.MutinyStub;
 import io.quarkus.grpc.RegisterClientInterceptor;
 import io.quarkus.grpc.RegisterInterceptor;
 import io.quarkus.grpc.RegisterInterceptors;
-import io.quarkus.grpc.runtime.MutinyBean;
-import io.quarkus.grpc.runtime.MutinyClient;
-import io.quarkus.grpc.runtime.MutinyGrpc;
-import io.quarkus.grpc.runtime.MutinyService;
-import io.quarkus.grpc.runtime.MutinyStub;
 import io.quarkus.grpc.runtime.supports.Channels;
 import io.quarkus.grpc.runtime.supports.GrpcClientConfigProvider;
 import io.smallrye.common.annotation.Blocking;
