+++ extensions/grpc/protoc/src/main/resources/MutinyBean.mustache
@@ -4,7 +4,7 @@ package {{packageName}};
 
 import io.grpc.BindableService;
 import io.quarkus.grpc.GrpcService;
-import io.quarkus.grpc.runtime.MutinyBean;
+import io.quarkus.grpc.MutinyBean;
 
 @javax.annotation.Generated(
 value = "by {{classPrefix}} Grpc generator",
