+++ extensions/grpc/protoc/src/main/resources/MutinyClient.mustache
@@ -4,7 +4,7 @@ package {{packageName}};
 
 import java.util.function.BiFunction;
 
-import io.quarkus.grpc.runtime.MutinyClient;
+import io.quarkus.grpc.MutinyClient;
 
 @javax.annotation.Generated(
 value = "by {{classPrefix}} Grpc generator",
