+++ extensions/grpc/protoc/src/main/resources/MutinyInterface.mustache
@@ -2,7 +2,7 @@
 package {{packageName}};
 {{/packageName}}
 
-import io.quarkus.grpc.runtime.MutinyService;
+import io.quarkus.grpc.MutinyService;
 
 @javax.annotation.Generated(
 value = "by {{classPrefix}} Grpc generator",
