+++ extensions/grpc/runtime/src/main/java/io/quarkus/grpc/runtime/devmode/GrpcServerReloader.java
@@ -9,8 +9,8 @@ import io.grpc.ServerMethodDefinition;
 import io.grpc.ServerServiceDefinition;
 import io.grpc.internal.ServerImpl;
 import io.quarkus.dev.testing.GrpcWebSocketProxy;
-import io.quarkus.grpc.runtime.ServerCalls;
-import io.quarkus.grpc.runtime.StreamCollector;
+import io.quarkus.grpc.stubs.ServerCalls;
+import io.quarkus.grpc.stubs.StreamCollector;
 import io.quarkus.runtime.LaunchMode;
 import io.quarkus.runtime.configuration.ProfileManager;
 import io.vertx.grpc.VertxServer;
