+++ extensions/grpc/stubs/src/main/java/io/quarkus/grpc/stubs/ClientCalls.java
@@ -1,4 +1,4 @@
-package io.quarkus.grpc.runtime;
+package io.quarkus.grpc.stubs;
 
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
