+++ extensions/grpc/stubs/src/main/java/io/quarkus/grpc/stubs/MultiStreamObserver.java
@@ -1,4 +1,4 @@
-package io.quarkus.grpc.runtime;
+package io.quarkus.grpc.stubs;
 
 import io.grpc.stub.StreamObserver;
 import io.smallrye.mutiny.subscription.MultiEmitter;
